user32 refactor, comments removed, require('electron') error suppressed
