+++ electron-quick-start/index.js
@@ -4,6 +4,6 @@ $('#setting-open-at-cursor').change(function cc() {
   ipc.send('setting-change-openAtCursor', this.checked);
 });
 
-$('#hide-to-tray-btn').click(function cc() {
+$('#hide-to-tray-btn').click(() => {
   ipc.send('hide-to-tray-btn');
 });
