+++ electron-quick-start/main.js
@@ -1,34 +1,10 @@
 const electron = require('electron');
 const { globalShortcut, ipcMain } = require('electron');
 const { clipboard } = require('electron');
-const ref = require('ref');
-const ffi = require('ffi');
 const { Tray, Menu } = require('electron');
+const User32Module = require('./lib/user32');
 
-const Struct = require('ref-struct');
-
-const MousePoint = Struct({
-  x: 'long',
-  y: 'long',
-});
-
-const MousePointPtr = ref.refType(MousePoint);
-
-const user32 = new ffi.Library('user32', {
-  GetCursorPos: ['long', [MousePointPtr, 'pointer']],
-  GetTopWindow: ['long', ['long']],
-  FindWindowA: ['long', ['string', 'string']],
-  SetActiveWindow: ['long', ['long']],
-  SetForegroundWindow: ['bool', ['long']],
-  BringWindowToTop: ['bool', ['long']],
-  ShowWindow: ['bool', ['long', 'int']],
-  SwitchToThisWindow: ['void', ['long', 'bool']],
-  GetForegroundWindow: ['long', []],
-  AttachThreadInput: ['bool', ['int', 'long', 'bool']],
-  GetWindowThreadProcessId: ['int', ['long', 'int']],
-  SetWindowPos: ['bool', ['long', 'long', 'int', 'int', 'int', 'int', 'uint']],
-  SetFocus: ['long', ['long']],
-});
+const user32 = new User32Module();
 
 let foregroundHWnd;
 let openAtCursor = false;
@@ -86,19 +62,11 @@ function createWindow() {
     hideToTray();
   });
 
-  // Open the DevTools.
-  // mainWindow.webContents.openDevTools()
-
-  // Emitted when the window is closed.
   mainWindow.on('closed', () => {
-    // Dereference the window object, usually you would store windows
-    // in an array if your app supports multi windows, this is the time
-    // when you should delete the corresponding element.
     mainWindow = null;
     childWindow.close();
   });
 }
-
 
 function createChildWindow() {
   childWindow = new BrowserWindow({
@@ -113,7 +81,6 @@ function createChildWindow() {
     slashes: true,
   }));
 
-  // Emitted when the window is closed.
   childWindow.on('closed', () => {
     childWindow = null;
   });
@@ -123,22 +90,15 @@ let childIsVisible = false;
 
 function toggleChildWindow() {
   if (childIsVisible) {
-    user32.SetForegroundWindow(foregroundHWnd);
-    user32.ShowWindow(foregroundHWnd, 5);
-    user32.SetFocus(foregroundHWnd);
-    user32.SetActiveWindow(foregroundHWnd);
+    user32.ReturnAllFocusToWindow(foregroundHWnd);
     childWindow.hide();
   } else {
     if (openAtCursor) {
-      const mousePosition = new MousePoint();
-      // Call user32 function GetCursorPos and Pass our MousePoint struct reference to it.
-      user32.GetCursorPos(mousePosition.ref(), null);
-      console.log(`X: ${mousePosition.x}, Y: ${mousePosition.y}`);
-      childWindow.setPosition(mousePosition.x, mousePosition.y);
+      const mouse = user32.GetCursorCoordinates();
+      childWindow.setPosition(mouse.x, mouse.y);
     } else {
       positioner.move('bottomRight');
     }
-
     childWindow.show();
     childWindow.focus();
   }
@@ -159,7 +119,6 @@ function clipboardListener() {
       console.log(nativeImage);
     },
     onTextChange: (text) => {
-      // console.log(text);
       childWindow.webContents.send('add-new-copy', text);
     },
   });
@@ -183,9 +142,6 @@ ipcMain.on('hide-to-tray-btn', () => {
   mainWindow.minimize();
 });
 
-// This method will be called when Electron has finished
-// initialization and is ready to create browser windows.
-// Some APIs can only be used after this event occurs.
 app.on('ready', () => {
   setHotkeys();
   clipboardListener();
@@ -194,23 +150,15 @@ app.on('ready', () => {
   createChildWindow();
 });
 
-// Quit when all windows are closed.
 app.on('window-all-closed', () => {
-  // On OS X it is common for applications and their menu bar
-  // to stay active until the user quits explicitly with Cmd + Q
-  if (process.platform !== 'darwin') {
+  if (process.platform !== 'darwin') { // OSX windows close to dock, not quit
     globalShortcut.unregisterAll();
     app.quit();
   }
 });
 
-app.on('activate', () => {
-  // On OS X it's common to re-create a window in the app when the
-  // dock icon is clicked and there are no other windows open.
+app.on('activate', () => { // OSX reactivate from dock
   if (mainWindow === null) {
     createWindow();
   }
 });
-
-// In this file you can include the rest of your app's specific main process
-// code. You can also put them in separate files and require them here.
